+++ Sources/CodexBar/StatusItemController+AgentSessions.swift
@@ -1,6 +1,31 @@
 import AppKit
 
 extension StatusItemController {
+    func wireAgentSessionUpdates() {
+        self.agentSessions.onUpdate = { [weak self] in
+            guard let self else { return }
+            if let latestActivityAt = self.agentSessions.latestLocalActivityAt {
+                self.store.noteCodingActivityObserved(at: latestActivityAt)
+            }
+            if self.settings.agentSessionsEnabled {
+                self.invalidateMenus(refreshOpenMenus: true)
+            }
+        }
+    }
+
+    func synchronizeAgentSessionsForSettingsChange() {
+        let remoteConfigurationChanged =
+            self.settings.agentSessionsEnabled != self.lastAgentSessionsEnabled ||
+            self.settings.agentSessionsManualHosts != self.lastAgentSessionsManualHosts
+        let monitoringChanged = self.settings.refreshFrequency != self.lastAgentSessionsRefreshFrequency
+        guard remoteConfigurationChanged || monitoringChanged else { return }
+
+        self.lastAgentSessionsEnabled = self.settings.agentSessionsEnabled
+        self.lastAgentSessionsManualHosts = self.settings.agentSessionsManualHosts
+        self.lastAgentSessionsRefreshFrequency = self.settings.refreshFrequency
+        self.agentSessions.settingsDidChange(remoteConfigurationChanged: remoteConfigurationChanged)
+    }
+
     @objc func focusAgentSession(_ sender: NSMenuItem) {
         guard let values = sender.representedObject as? [String],
               let sessionID = values.first